+++ src/main/java/duke/command/AddCommand.java
@@ -1,0 +1,4 @@
+package duke.command;
+
+public class AddCommand {
+}
