+++ src/main/java/duke/task/Deadline.java
@@ -1,7 +1,9 @@
 package duke.task;
 
-import duke.task.Task;
 import duke.type.TaskType;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.time.temporal.ChronoUnit;
 
 public class Deadline extends Task {
     protected String by;
